新增：搜尋欄元件化
